cleaned up code; changed all vec3 to dvec3

+++ project/Light.hpp
@@ -8,8 +8,8 @@
 struct Light {
     Light();
 
-    glm::vec3 colour;
-    glm::vec3 position;
+    glm::dvec3 colour;
+    glm::dvec3 position;
     double falloff[3];
 };
